Added login button

+++ src/components/Admin/DealershipGroups/DealershipGroups.tsx
@@ -1,7 +1,7 @@
-import React from "react";
+import React, {useState} from "react";
 import {Table} from "../../UI/Table";
-import {IconButton} from "@material-ui/core";
-import {DeleteForever, Visibility} from "@material-ui/icons";
+import {IconButton, Menu, MenuItem, Typography} from "@material-ui/core";
+import {MoreHoriz} from "@material-ui/icons";
 import {TableRowDataType} from "../../UI/types";
 import {IDealershipGroupExtended} from "../../../store/reducers/dealershipGroups/types";
 import {useDispatch, useSelector} from "react-redux";
@@ -36,35 +36,57 @@ export const DealershipGroups = () => {
         changePageData
     );
 
+    const [editedItem, setEditedItem] = useState<IDealershipGroupExtended|null>(null);
+    const [anchorEl, setAnchorEl] = useState<EventTarget&HTMLButtonElement|null>(null);
+
     const {askConfirm} = useConfirm();
     const dispatch = useDispatch();
     const showError = useException();
     const showMessage = useMessage();
     const history = useHistory();
 
-    const handleView = (el: IDealershipGroupExtended) => () => {
-        history.push(`${Routes.Admin.DealershipGroups}/${el.id}`);
+    const handleView = ()  => {
+        if (!editedItem) return;
+        setAnchorEl(null);
+        history.push(`${Routes.Admin.DealershipGroups}/${editedItem.id}`);
     };
-    const handleRemoveAction = (el: IDealershipGroupExtended) => () => askConfirm({
-        title: `Are you sure want to remove dealership group ${el.name}?`,
-        isRemove: true,
-        onConfirm: async () => {
-            await handleRemove(el);
-        }
-    });
-    const handleRemove = async (d: IDealershipGroupExtended) => {
+    const handleRemoveAction = () => {
+        setAnchorEl(null);
+        askConfirm({
+            title: `Are you sure want to remove dealership group ${editedItem?.name}?`,
+            isRemove: true,
+            onConfirm: async () => {
+                await handleRemove();
+            }
+        });
+    };
+    const handleRemove = async () => {
+        if (!editedItem) return;
+        const d = editedItem
         try {
             await dispatch(removeDealership(d.id));
             showMessage(`Successfully removed ${d.name}`);
         } catch (e) {
             showError(e);
         }
-
     }
-    const viewActions = (el: IDealershipGroupExtended) => (<>
-        <IconButton size="small" onClick={handleView(el)}><Visibility /></IconButton>
-        <IconButton size="small" style={{marginLeft: 8}} onClick={handleRemoveAction(el)}><DeleteForever /></IconButton>
-    </>);
+
+    const closeMenu = () => {
+        setEditedItem(null);
+        setAnchorEl(null);
+    }
+
+    const openMenu = (el: IDealershipGroupExtended) =>
+        (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+        setEditedItem(el);
+        setAnchorEl(e.currentTarget);
+    }
+
+    const viewActions = (el: IDealershipGroupExtended) =>
+        <IconButton size="small" onClick={openMenu(el)}>
+            <MoreHoriz />
+        </IconButton>
+    ;
 
     React.useEffect(() => {
         dispatch(dealershipActions.loadAll());
@@ -85,5 +107,12 @@ export const DealershipGroups = () => {
             index="id"
             actions={viewActions}
         />
+        <Menu open={Boolean(anchorEl)} anchorEl={anchorEl} onClose={closeMenu}>
+            <MenuItem onClick={handleView}>View</MenuItem>
+            <MenuItem>Login</MenuItem>
+            <MenuItem onClick={handleRemoveAction}>
+                <Typography color="secondary">Remove</Typography>
+            </MenuItem>
+        </Menu>
     </>
 }
